docs(readme): don't call it a prototype any more
The console has now reached a state where it's more than just a
_prototype_ --- we're actively developing the software into a
releaseable state. I think it's still important to make it clear that
this software isn't stable and is in a pre-alpha state, but it's no
longer just a prototype.

I've edited the language in the readme a little bit to reflect this.