fix(subscriber): fix potential spurious flush notifications
Currently, there is a potential issue with the flush notification
channel. An `AtomicBool` is used to determine whether a flush
notification should be sent, so that only one thread will trigger a
flush when approaching the flush capacity. However, the `AtomicBool` is
cleared by the aggregator task _as soon as_ it's woken by the flush
notification. This seems to potentially cause a situation where several
notifications are sent while the channel capacity is below the flush
limit, because the flag is cleared as soon as the aggregator task is
_woken_, and before it has the time to actually drain the channel. This
seems to result in many notifications being queued, causing the
aggregator task to loop a bunch of times.

This branch fixes this by changing the aggregator task to only clear the
flag when it has drained all queued events, hopefully reducing needless
notifications. This appears to partially resolve #177, although there
appears to also be an issue where console *updates* seem to become
increasingly delayed over time.