docs(subscriber): use H1 headers in builder API docs (#204)
This changes the `Builder` API docs to use H1 (`#`) markdown headers
rather than H2 (`##`) headers for top-level sections in the
documentation. It turns out that when the docs are on a
function/method, RustDoc makes the H2 headers really small and it looks
weird.

**Before**:
![image](https://user-images.githubusercontent.com/2796466/146258971-27d08d26-f862-4185-838c-526b46e3ac96.png)
Note how weirdly tiny the header text is.

**After**:
![image](https://user-images.githubusercontent.com/2796466/146259132-1af30df4-7196-418c-91f5-ea2a1158ead1.png)
Now, the header text is a more normal size!