feat(subscriber): rename `TasksLayer` to `ConsoleLayer`
The `TasksLayer` type doesn't *just* record data from task spans; it
also records async op and resource spans and events.

When it was initially added, it only recorded task data, and it was named
`TasksLayer` because I wasn't sure if we were eventually going to add
other layers for other kinds of data or not. Now that it also collects
the async op and resource spans, we should maybe give it a name that
doesn't suggest it's _specifically_ for tasks.

This branch renames `TasksLayer` to `ConsoleLayer`.

The other option would be to just call it `console_subscriber::Layer`.
But, I'm not as big a fan of that, as it means that if the
`tracing_subscriber::Layer` trait is also in scope, the
`console_subscriber::Layer` type would always have to be prefixed; it
can't be imported. So, I thought `ConsoleLayer` was better.