fix(console): sort attributes alphabetically by name (#210)
Currently, the attributes for async ops are not sorted before displaying
them, so the same attribute names may occur in different orders. This is
"not great".

This branch fixes this by sorting the `Attribute`s prior to formatting
them (in `Attribute::make_formatted`), the same way we do in
`Field::make_formatted`. I changed the custom comparison closure into an
`Ord` impl for `Field` so it could also be used in an `Ord` impl for
`Attribute`.

As you can see, it works: 
![image](https://user-images.githubusercontent.com/2796466/146440281-5cefca2a-75bf-470d-82e5-cdbccc967ed3.png)

Fixes #208