fix(subscriber): ignore spans that weren't initially recorded
Currently, a strange behavior exists in the `console-subscriber`
crate. If a span for a task, async op, or resource is created, and the
event buffer is full, the aggregator task will not be informed of that
span's creation. But, if the event buffer then empties out, we might
send the aggregator task enter/exit/close events for that span. This
results in the aggregator receiving events for an unknown span, which
might result in panics or subtly wrong data.

This branch fixes this by changing the `ConsoleLayer` to track whether
it was able to successfully send an event for the creation of a span,
and to only care about subsequent events on that span if it successfully
recorded the span's creation. This is done by inserting a zero-sized
marker type into the span's extensions map, and ignoring spans that
don't have this marker.