chore: just automate binary publishing for now
Looks like Taiki's action can't currently handle tags with prefixes,
so they have to be _just_ version strings (see
taiki-e/create-gh-release-action#1).

So, we can consider automatically generating the release when it
supports tags of the format we're using (separate prefixes for each
workspace crate). For now, let's just automate the building and
uploading of release binaries, and manually create the releases...