fix(console): don't enable crossterm mouse capture
Currently, the console captures mouse input, but it doesn't actually do
anything with mouse input yet. This is a shame, because it means you
can't do things like copying and pasting text out of the console.

This commit disables mouse capture.

In the future, if we add mouse support to the UI, we can re-enable mouse
capture...but it would be nice to make sure copying and pasting still
works (we may have to reimplement it?).

Fixes #167