fix(console): increase default event buffer capacity a bit
It seems like a number of people using the console with real life
applications are running into issues related to the console's event
buffer being at capacity (e.g. #230, #234, etc). This commit bumps up
the default capacity a bit as a quick fix.

In the future, I think we should probably do things to avoid reaching
the event buffer capacity as often --- I think we can change the
aggregator/layer design a bit to avoid creating as many events. But, for
now, this should hopefully give people a better default configuration.