Display help popup correctly.
We want to display the help popop in the current view, no matter
whichever it might be. This is because we might have distinct help
infomation per view. Therefore, the help view being displayed is
unrelated to any state in which the console as a whole is.

We handle this by separating the rendering of the help popup from the
state of the application.