chore(console): rename `console` dir to `tokio-console` (#263)
Currently, the `tokio-console` CLI's crate is named `tokio-console`, but
the directory is just named `console`. This causes some problems: for
example, when we publish a release, the release tags are named
`tokio-console-vX.Y.Z`. The GitHub Action for automatically publishing
changelogs to GitHub Releases looks for the changelog in
`<tag-prefix>/CHANGELOG.md`, but this doesn't work, because the tag
prefix is not the name of the directory. So, the automatic release
publishing currently fails for the console crate:
https://github.com/tokio-rs/console/runs/4860467681?check_suite_focus=true

I figured the easiest solution was just to rename the directory to match
the crate name --- this is more consistent anyway.