chore(api): move timestamp proto file (#277)
This PR moves
`console-api/proto/google/protobuf/timestamp/timestamp.proto` to
`console-api/proto/google/protobuf/timestamp.proto` to match the
directory structure in https://github.com/protocolbuffers/protobuf. This
may not be important for generating code for Rust, however other
languages such as Javascript need the path to match the official
definitions in order to work with the official
[google-protobuf](https://www.npmjs.com/package/google-protobuf)
library.