fix(subscriber): fix compilation on targets without 64-bit atomics (#282)
## Motivation

Currently, the `console-subscriber` crate does not compile on targets
without 64-bit atomic operations, such as `mipsel-unknown-linux-musl`.
This is because a 64-bit atomic is used to store the last task ID that
polled a resource.

## Solution

This branch changes this code to use the `AtomicCell` type from
`crossbeam-utils`, instead. `AtomicCell` will use atomic operations on
targets that support 64-bit atomics, and fall back to using a mutex on
platforms that don't have 64-bit atomic operations. This should fix the
build on those platforms.

Fixes #279