docs: add information on where to put .cargo/config.toml (#284)
I ran into an issue where putting my .cargo/config.toml with the
tokio_unstable rustflag in the sub-project of my workspace where I
needed to use tokio-console led to full rebuilds on every cargo build
invocation.

This was caused by rust-analyzer invoking its cargo commands in the
workspace root independent of the directory where the editor (e.g. nvim)
is launched from. This led to rust-analyzer not to use the cargo config,
while my manual cargo build invocations did use the config, hence
requiring full rebuilds on every run.

This documentation informs new users of this pitfall.

If this documentation is found to be a little on the long side, I can
try to move part of this warning into the cargo documentation, which I'm
linking to anyways.