style(console): fix Clippy lints on "unnecessary" `to_owned` calls (#302)
Currently, Clippy's [`unnecessary_to_owned`][1] lint is triggering in a
couple of places in `tokio-console`.

In this case, the `to_owned` calls actually _are_ necessary before
passing a string to `Cell::new` in a table view. However, in a couple
places, we call `to_owned` _inside_ of the argument to
`Width::update_str`, which does _not_ require an owned string, so Clippy
thinks the `to_owned` call is unnecessary. `update_str` returns the
argument, which is then passed to `Cell::new`.

This commit moves the `to_owned` call to _after_ the `update_str` call,
so that the string is passed to `update_str` as an `&str`, and
`to_owned` is only called afterwards. This doesn't actually meaningfully
change the code in any way, but it makes Clippy happy. :)

[1]: https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_to_owned