chore(CI): don't trigger duplicate CI runs on pushes
Currently, CI is configured to run on all pushes, and on all pull
requests. This means that when changes are pushed to a PR, the CI jobs
run *twice*. That's not great, as it means our builds take a bit longer.

This branch fixes the CI configuration, so that CI jobs are triggered on
all pull requests, and on pushes to `main`, rather than on all pushes.
Now, all jobs should only run a single time for every push to a PR.