chore: add issue templates and security policy (#304)
Currently, when people report issues with the console, they just get a
blank issue. This makes it difficult to ensure that we get consistent
information reported for issues and suggestions.

This branch adds GitHub issue templates for bug reports and feature
requests, which should generate a form people can fill in with specific
information. Also, I added the security policy from the Tokio
repository, with a link for reporting security vulnerabilities, because
the console should follow the same security policy as the rest of Tokio.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>