Don't trace tasks spawned through the console subscriber by default
Enabling the console subscriber starts a server that generates a lot of
async activity.  Introspecting or logging that can drown out the
activity the program has outside of instrumentation.

Provide a new builder method, enable_self_trace, to control whether to
also trace events that come with enabling the tokio console.

If self trace is not enabled, set a thread-local subscriber that drops
events coming from the server thread.

Raise the requirement on tracing-core so that setting a per-thread
NoSubscriber is effective.