console-api: Update `tonic` to `0.`7
This change updates `tonic` and `prost` to their latest versions. In
addition this PR changes how the generated protobuf is built from being
done in a `build.rs` file to a `tests/boostrap.rs` test. This removes
the need for downstream consumers to compile/provide `protoc` for
`prost` to generate the protobuf files (which was changed in `0.10`).

The bootstrap test will first attempt to compile the protobuf files and
check if there is a git diff between the checked in version. If there is
a difference it will fail the test waiting for the files to be checked
in. This will fail PRs that have updated the protobuf files but not
checked in the new versions.