docs: add tokio-console installation section (#313)
## Description

This PR adds a section for installing the tokio-console crate as well as
fixes some broken console links. All changes were made to the README.
Even though the tokio-console crate has its own README that includes
installation, users will likely stumble upon the root README first.
Thus, the purpose of this PR is to aid folks who may not wish to clone
the repository when trying out the project.

## Commit Notes

- Add tokio-console installation section in the README as the primary
  method of using tokio-console
- Move the local cargo run option to its own alternative method section
- Update broken console links to use tokio-console in the README
- Add ".idea" to gitignore
- Update cargo install to use "--locked" flag in the nested README

## Additional Notes

Excellent project! Thank you for this.

Signed-off-by: Nick Gerace <nickagerace@gmail.com>
Co-authored-by: Eliza Weisman <eliza@buoyant.io>