feat(console): add autogenerated example config file to docs (#327)
This branch updates `tokio-console`'s documentation to include a
configuration reference describing the config file and the CLI
arguments.

In addition, I've added code to automatically generate the example
config file and the help text for the CLI arguments as files that can be
checked in to Git. The example text is generated in a test, which will
fail if the generated output has changed. This way, we can ensure that
changes that change CLI arguments or config file fields will include
updates to the example files. In the RustDoc configuration reference, we
use `include_str!` to pull in those files. This means the docs are
always kept up to date, and the actual code that defines the config file
and the CLI args are used as a single source of truth from which the
documentation can be generated.

Check out the [Netlify deploy preview][1] to see the new docs!

[1]: https://62570bc00775d9000823486d--tokio-console.netlify.app/tokio_console/config_reference/index.html

Signed-off-by: Eliza Weisman <eliza@buoyant.io>